Changed the way guns work for it to fit the project guidlines
Made sure the Pistol works, it now has a raycast instead of projectiles.The ennemies dies after 3 hits of the gun. The sword now has an animation that will instantly kill any enemies the sword collider enters in contact with, this goes with the enemy scripts that gives them health and will disappear if they get their hp lowered to 0. Also added a script that switches in between the Pistol and and the sword wihch displayed in the ui as well as the ammo for the pistol